Improve UX by adding Pop-up feature to Artwork Gallery
This commit incorporates a new Pop-up feature to the Artwork Gallery which will provide users with more detailed information about each artwork in a separate pop-up window. This was achieved by importing a `ModalComponent` renamed to `PopinComponent` in the `GalleryView.vue`. A `togglePopin` function was also added. As part of the UX improvement, animations have been added to the pop-up, artwork title and description. Also, the original button for `toggleModal` was updated to trigger the new `togglePopin` function.