"Added Popin functionality and updated SCSS color variables
The commit introduces the ‘Popin’ functionality to the Gallery component, which provides a detailed description for artworks when a user selects 'En savoir plus'. This has been implemented as a section nested within the Gallery component, and is responsible for presenting information about the artworks as a popin box. Previously, the 'En savoir plus' button had no functionality.
The node dependencies for the Express.js backend were also updated from 20.8.9 to 20.10.4 in the package.json and package-lock.json files to ensure the latest dependencies are being used.
A color variable for red was also added to the main SCSS file to help maintain consistency when using this color in the CSS."