"Added artists JSON file to Express API database
This commit introduces a new JSON file 'artists.json' containing the data for the artists in the database. This was necessary to allow the Express API to fetch and manage artist data independently from the rest of the application. Now we will have data at our disposal for functionalities regarding the display, addition, update or deletion of artist information."