Adjust style properties and update README.md
This commit adjusts various style properties in ArtistsView.vue and GalleryView.vue files to improve the visual hierarchy and readability of the website. The font size of the item description has been lowered to 1.7rem to prevent overwhelming the page. A new CSS class, '__byletter', has been included to provide a better presentation for the artists by letter section with the subtle addition of a background color and border. These changes were made following the UI/UX best practices for improved user engagement. Besides, a preview of the gallery and artists view has been added to the README.md to provide a visual reference for users and developers.